ec2/requirements.txt
mysql-connector-python
boto3
requests
 
